Improve: Rename specification files, update
 - Update main spec to match spec (v0.0.16)
 - There will need to be further updates to match spec
